Move all pages and related components to dedicated directories
